+++ bin/actions/bundle_archive_action.js
@@ -3,11 +3,13 @@ const StatusLog = require('../support/status_log')
 const bundleArchiveAction = (resourceBundle, manifestConfiguration) => {
   const { archiveBundle, PackageTool } = resourceBundle
 
-  StatusLog.notify(`bundle ${archiveBundle.uri}`, archiveBundle.uuid)
+
   return PackageTool
     .pack(archiveBundle, manifestConfiguration)
     .then(({ releaseAsset }) => {
+      StatusLog.notify(`packaged`, archiveBundle.uuid)
       resourceBundle.archiveBundle.releaseAsset = releaseAsset
+
       return resourceBundle
     })
 }