+++ bin/commands/download.js
@@ -18,19 +18,15 @@ module.exports = {
             const archiveDependencies = archiveConfiguration.allDependencies()
             const archiveName = archiveConfiguration.name
 
-            StatusLog
-              .initialize()
-              .start()
-
             downloadArchivesAction(archiveDependencies, archiveName)
-              .catch(err => {
-                StatusLog
-                  .completeFailure(err.toString())
-                  .then(() => done())
-              })
               .then(() => {
                 return StatusLog
                   .completeSuccess()
+                  .then(() => done())
+              })
+              .catch(err => {
+                StatusLog
+                  .completeFailure(err.toString())
                   .then(() => done())
               })
           })