+++ bin/support/status_log.js
@@ -1,34 +1,19 @@
 const winston = require('winston')
 const color = require('cli-color')
-const process = require('process')
 const _ = require('lodash')
 
-const Colors = {
-  gray: color.xterm(8),
+const C = {
+  gray: color.xterm(249),
+  darkGray: color.xterm(242),
+  lightGray: color.xterm(253),
   yellow: color.yellow,
-  red: color.redBright
+  red: color.redBright,
+  action: color.xterm(139),
+  resource: color.xterm(145),
+  error: color.xterm(198)
 }
 
 class StatusLog {
-
-  static __drawContent () {
-    const spinner = Spinner.running[this.__frame % Spinner.running.length]
-    const seconds = parseInt(this.__frame / (1000 / this.__refreshRate), 10)
-
-    const errorCount = _.size(this.__errors)
-    const errors = (errorCount > 0) ? Colors.red(errorCount + '! ') : ``
-    const action = ``
-
-    return `${Colors.yellow(spinner)} ${Colors.gray(seconds + 's')} ${errors}${action}`
-  }
-
-  static start () {
-    return this
-  }
-
-  static stop () {
-    return this
-  }
 
   static initialize (vorpal) {
     this.uninitialize()
@@ -40,7 +25,7 @@ class StatusLog {
         timestamp: () => Date.now(),
         handleExceptions: true,
         filename: `vault.log`,
-        json: false,
+        json: true,
         options: {
           flags: 'w'
         }
@@ -60,20 +45,42 @@ class StatusLog {
     this.__logger = undefined
     this.__vorpal = undefined
     this.__errors = []
-    this.__frame = 0
+    this.__started = Date.now()
+    this.__messageColors = [
+      C.action,
+      C.resource,
+      C.error
+    ]
 
     return this
   }
 
   static notify (action, resource, meta = {}) {
-    this.__logger.info(`[${resource}] ${action}`, meta)
-    this.__vorpal.log(action)
+    const seconds = `${(Date.now() - this.__started) / 1000}`.slice(0, 3) + `s`
+    const secondsStr = (seconds > 10) ? `${seconds} ` : `${seconds}  `
+    const colors = this.__messageColors
 
+    const errorCount = this.__errors.length
+    const errors = (errorCount > 0) ? errorCount + '!' : ``
+
+    const message = [
+      action,
+      resource,
+      errors
+    ]
+
+    this.__logger.info(message.join(` `), meta)
+    this.__vorpal.log(`${C.darkGray(secondsStr)}` + message.map((s, i) => colors[i](s)).join(` `))
     return this
   }
 
   static error (reason, resource, meta = {}) {
-    this.__errors.push(reason)
+    this.__errors.push({
+      reason,
+      resource,
+      meta
+    })
+
     return this
   }
 
@@ -81,7 +88,7 @@ class StatusLog {
     this.notify(`completed`)
 
     return new Promise((resolve, reject) => {
-      resolve(this.stop().__writeErrors().uninitialize())
+      resolve(this.__writeErrors().uninitialize())
     })
   }
 
@@ -90,17 +97,18 @@ class StatusLog {
 
     StatusLog.error(reason)
     return new Promise((resolve, reject) => {
-      resolve(this.stop().__writeErrors().uninitialize())
+      resolve(this.__writeErrors().uninitialize())
     })
   }
 
   static __writeErrors () {
-    const errors = _.reduce(this.__errors, (content, reason, index) => {
-      return `${content}${index + 1}. ${Colors.gray(reason)}\n`
+    const errors = _.reduce(this.__errors, (content, { reason, resource }, index) => {
+      return `${content}${index + 1}. ${C.gray(reason)}\n`
     }, ``)
 
     if (!_.isEmpty(errors)) {
       this.__logger.error(errors)
+      this.__vorpal.log(errors)
     }
 
     return this