+++ bin/transits/ftp_transit.js
@@ -1,4 +1,3 @@
-const applicationConfiguration = require('../configurations/application')
 const SecurityServant = require('../servants/security_servant')
 const FileSystem = require('../support/file_system')
 const path = require('path')
@@ -6,7 +5,7 @@ const _ = require('lodash')
 const FTP = require('ftp')
 
 class FTPTransit {
-  static push(archiveBundle) {
+  static push (archiveBundle) {
     return new Promise((resolve, reject) => {
       const pushFiles = FileSystem.flattenFolder(archiveBundle.releaseAsset)
       const ftpInfo = this.__ftpInfo(archiveBundle)
@@ -18,22 +17,19 @@ class FTPTransit {
       })
 
       ftp.on('ready', () => {
-        const sourceScope = archiveBundle.sourceScope
         let uploaded = pushFiles.length
 
         pushFiles.forEach((file) => {
           const aspects = file.split(path.sep)
           const folder = _.initial(aspects).join(path.sep)
-          const filename = _.last(aspects)
 
           ftp.mkdir(folder, recursivelyCreate, () => {
             ftp.put(file, ftpInfo.filePath, (err) => {
               uploaded -= 1
 
               if (err) {
-                reject(new Error(err.toString()))
-              }
-              else if (uploaded <= 0) {
+                reject(new Error(err))
+              } else if (uploaded <= 0) {
                 ftp.end()
                 resolve()
               }
@@ -46,16 +42,16 @@ class FTPTransit {
     })
   }
 
-  static pull(archiveRequest) {
+  static pull (archiveRequest) {
     return new Promise((resolve, reject) => {
-      reject('')
+      reject(new Error(`not implemented (ftp::pull)`))
     })
   }
 
-  static __ftpInfo({ uri, scope, scopeSource }) {
+  static __ftpInfo ({ uri, scope, scopeSource }) {
     const PARSE_FTP = /(ftp:\/\/)([a-z0-9A-Z]+?)@([.a-z0-9A-Z]+?):([0-9]+?)\/(.*)/
     const ftpURI = PARSE_FTP.exec(uri)
-    console.dir(ftpURI)
+
     const ftpInfo = {
       filePath: `/${ftpURI[5]}`,
       host: ftpURI[3],
@@ -70,7 +66,6 @@ class FTPTransit {
     ftpInfo.password = SecurityServant.decrypt(encryptedKey)
     return ftpInfo
   }
-
 }
 
 module.exports = FTPTransit