+++ changelog.md
@@ -1,5 +1,12 @@
 # Klepto Changelog
 
+## 0.9.0
+
+- ADDED documentation
+- ADDED `configuration` will show the entire configuration object affecting a project folder
+- FIXED always include the local folder last and the global folder before the folder traversal when calculating the application configuration
+- FIXED make sure the global configuration is before any folder traversal configurations
+
 ## 0.8.0
 
 - ADDED `klepto publish` to prepare a build an archive and upload to a host